requirements.txt
fastapi
uvicorn
python-multipart
requests
pillow
transformers
torch
